+++ src/utils.c
@@ -668,10 +668,11 @@ push_env(char *str, char ***env)
 
 /* Display final summary */
 void
-display_final_summary(const fsize_t total_size, const fnum_t total_num_files)
+display_final_summary(const pnum_t total_num_parts, const fsize_t total_size,
+    const fnum_t total_num_files)
 {
-    fprintf(stderr, "Total: size = %ju, files = %ju\n",
-        total_size, total_num_files);
+    fprintf(stderr, "Summary: size = %ju, files = %ju, parts = %ju\n",
+        total_size, total_num_files, total_num_parts);
 
     return;
 }