+++ docs/www.fpart.org/docs/changelog.md
@@ -17,6 +17,8 @@
     - fpsync: show transmitted files count / data size in final status
       and SIGINFO handler (fixes GH issue #24)
     - fpart: add two new hook variables: FPART_TOTALSIZE and FPART_TOTALNUMFILES
+    - fpart: final summary now shows total number of parts created, as well as
+      total number of files packed and total resulting size
 ```
 # Current: 1.5.1
 ```nohighlight